requirements.txt
asgiref==3.2.7
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
daphne==2.5.0
DateTime==4.3
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-guardian==2.2.0
django-jalali==4.0.0
django-jalali-date==0.3.1
django-rest==0.8.6
django-rest-auth==0.9.5
django-rest-registration==0.5.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
ippanel==1.1.0
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
nexmo==2.4.0
oauthlib==3.1.0
Pillow==7.1.2
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
twilio==6.43.0
Twisted==20.3.0
txaio==20.4.1
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
zope.interface==5.1.0
